Make authz ctx aware of peer name
When reading imported data we want to require different permissions from
those that would be granted for local services.

By adding the peer name of the requested resource to the authorizer
context we can then have authorizers implement different logic for data
imported from a peer.

This commit also updated how we fill the authz context, so that
resources that can be imported will fill in the peer name.